requirements.txt
requests==2.28.2
PyYAML==6.0
six==1.16.0
google-cloud==0.34.0
google-auth==2.16.0